Make ExceptionHandler more robust
If the system fell over due to a syntax error or some other issue before a ServerRequest object was bound into the container, a weird error message would be shown about the DI container. This was because the request couldn’t be retrieved from the container but totally obfuscated the root cause of the issue.