Add view model support
Add the ViewModel baseclass and also augment the `TimberResponse` to auto flatten the passed in context.

The processing recursively supports flattening instances of `ViewModel` and `Collection`.

Flatten context based on `Arrayable` interface

Rather than using the concrete `ViewModel` class, we have abstracted the behaviour to an `Arrayable` interface that can be added to any object that needs to be flattened in the context.

We’ve also moved to using the `Arrayable` interface from the Collections package instead of the concrete `Collection` class. This enables other objects that implement this to also be flattened.