Prevent duplicate headers being sent
This was an issue experienced on MAMP PRO where duplicate headers in the response was causing an Internal Server Error.

Unfortunately, it’s not possible to put tests around this as `headers_list()` does not work in the CLI. We may need to catch this at an integration level in the future.