Fix dependency issue
There seems to be an issue with older versions of `antecedent/patchwork` which are not present in a more recent version. Although this isn’t a direct dependency of ours we can increase the minimum version (to the latest).

Hopefully we can remove this in time if Brain Monkey increases it’s minimum accepted version.